verification_system/requirements-gateway.txt
redis>=4.5.0
websockets>=11.0.0
PyJWT>=2.8.0
cryptography>=41.0.0